Add scripts for checking code formatting
This commit use the git pre-commit hook to check and report the code
formatting problem in the committed codes.

However, as the directory .git/hooks cannot be tracked by git,
additional configuration is required, which is written in the README
file.

The reason why not merge the scripts into .githooks/pre-commit is for
future expansion (e.g., compilation check and reuse the script for
GitHub Action).